docs: update contributions and delete notes/log
I decided to get rid of the notes thing as I'm not really planning to
use it and it just seems like it's going to be too much of a a pain to
manage. I don't want to add extra tech debt that I'm never going to
touch.

I also added the contributions to include the new PR I did :).